+++ src/dso/_util.py
@@ -1,19 +1,30 @@
+from __future__ import annotations
+
 import importlib
 import json
 import subprocess
 import sys
 from collections.abc import Sequence
+from contextlib import contextmanager
 from functools import cache
 from importlib import resources
-from importlib.abc import Traversable
+from os import environ
 from pathlib import Path
-from typing import Literal
+from typing import TYPE_CHECKING, Literal
 
-from git.repo import Repo
-from jinja2 import StrictUndefined, Template
 from rich.prompt import Confirm
 
 from dso._logging import console, log
+
+# compatibilty with python 3.10
+try:
+    import tomllib
+except ImportError:
+    import tomli as tomllib
+
+
+if TYPE_CHECKING:
+    from importlib.resources.abc import Traversable
 
 DEFAULT_BRANCH = "master"
 
@@ -31,7 +42,7 @@ def check_project_roots(paths: Sequence[Path]) -> Path:
     return tmp_project_roots.pop()
 
 
-def _find_in_parent(start_directory: Path, file_or_folder: str, recurse_barrier: Path | None = None) -> Path | None:
+def find_in_parent(start_directory: Path, file_or_folder: str, recurse_barrier: Path | None = None) -> Path | None:
     """
     Recursively walk up to the folder directory until we either find `file_or_folder` or reach the root.
 
@@ -42,7 +53,9 @@ def _find_in_parent(start_directory: Path, file_or_folder: str, recurse_barrier:
     If the root is reached without finding the file, None is returned.
     """
     return _find_in_parent_abs(
-        start_directory.absolute(), file_or_folder, recurse_barrier.absolute() if recurse_barrier is not None else None
+        start_directory.absolute(),
+        file_or_folder,
+        recurse_barrier.absolute() if recurse_barrier is not None else None,
     )
 
 
@@ -86,7 +99,7 @@ def get_project_root(start_directory: Path) -> Path:
     FileNotFoundError
         If the .git folder is not found.
     """
-    proj_root = _find_in_parent(start_directory, ".git")
+    proj_root = find_in_parent(start_directory, ".git")
     if proj_root is None:
         raise FileNotFoundError("Not within a dso project (No .git directory found)")
     else:
@@ -94,13 +107,15 @@ def get_project_root(start_directory: Path) -> Path:
         return proj_root.parent
 
 
-def _get_template_path(template_type: Literal["init", "folder", "stage"], template_name: str) -> Traversable:
+def get_template_path(template_type: Literal["init", "folder", "stage"], template_name: str) -> Traversable:
     template_module = importlib.import_module(f"dso.templates.{template_type}")
     return resources.files(template_module) / template_name
 
 
 def _copy_with_render(source: Traversable, destination: Path, params: dict):
     """Fill all placeholders in a file with jinja2 and save file to destination"""
+    from jinja2 import StrictUndefined, Template
+
     with source.open() as f:
         template = Template(f.read(), undefined=StrictUndefined)
     rendered_content = template.render(params)
@@ -111,17 +126,21 @@ def _copy_with_render(source: Traversable, destination: Path, params: dict):
             file.write("\n")
 
 
-def _instantiate_template(template_path: Traversable, target_dir: Path | str, **params) -> None:
+def instantiate_template(template_path: Traversable, target_dir: Path | str, **params) -> None:
     """Copy a template folder to a target directory, filling all placeholder values."""
+    from jinja2 import Template
+
     target_dir = Path(target_dir)
 
     def _traverse_template(curr_path, subdir):
         for p in curr_path.iterdir():
             if p.is_file():
                 name_rendered = Template(p.name).render(params)
-                target_file = target_dir / subdir / name_rendered
-                if not target_file.exists():
-                    _copy_with_render(p, target_file, params)
+                # this file is used for checking in empty folders in git.
+                if name_rendered != ".gitkeeptemplate":
+                    target_file = target_dir / subdir / name_rendered
+                    if not target_file.exists():
+                        _copy_with_render(p, target_file, params)
             else:
                 (target_dir / subdir / p.name).mkdir(exist_ok=True)
                 _traverse_template(p, subdir / p.name)
@@ -129,16 +148,18 @@ def _instantiate_template(template_path: Traversable, target_dir: Path | str, **
     _traverse_template(template_path, Path("."))
 
 
-def _instantiate_with_repo(template: Traversable, target_dir: Path | str, **params) -> None:
+def instantiate_with_repo(template: Traversable, target_dir: Path | str, **params) -> None:
     """Create a git repo in a directory and render a template inside.
 
     Creates an initial commit.
     """
+    from git.repo import Repo
+
     target_dir = Path(target_dir)
     target_dir.mkdir(exist_ok=True)
     log.info("Created project directory.")
 
-    _instantiate_template(template, target_dir, **params)
+    instantiate_template(template, target_dir, **params)
     log.info("Created folder structure from template.")
 
     if not (target_dir / ".git").exists():
@@ -152,7 +173,7 @@ def _instantiate_with_repo(template: Traversable, target_dir: Path | str, **para
         log.info("Initalized local git repo.")
 
 
-def _git_list_files(dir: Path) -> list[Path]:
+def git_list_files(dir: Path) -> list[Path]:
     """
     Recursively list all files in `dir` that are not .gitignored.
 
@@ -160,11 +181,30 @@ def _git_list_files(dir: Path) -> list[Path]:
     Source: https://stackoverflow.com/a/77197460/2340703
     """
     res = subprocess.run(
-        ["git", "ls-files", "--cached", "--others", "--exclude-standard"], cwd=dir, capture_output=True
+        ["git", "ls-files", "--cached", "--others", "--exclude-standard"],
+        cwd=dir,
+        capture_output=True,
     )
     if res.returncode:
         sys.exit(res.returncode)
     return [dir / Path(p) for p in res.stdout.decode("utf-8").strip().split("\n")]
+
+
+@cache
+def get_dso_config_from_pyproject_toml(dir: Path) -> dict:
+    """
+    Read the pyproject.toml file when within a project and return the [tool.dso] section as dict
+
+    If the pyproject.toml file doesn't exist, or it doesn't have a [tool.dso] section,
+    an empty dictionary is returned
+    """
+    project_root = get_project_root(dir)
+    pyproject_toml = project_root / "pyproject.toml"
+    if not pyproject_toml.exists():
+        return {}
+    with pyproject_toml.open("rb") as f:
+        data = tomllib.load(f)
+    return data.get("tool", {}).get("dso", {})
 
 
 def _read_dot_dso_json(dir: Path):
@@ -200,12 +240,32 @@ def _update_dot_dso_json(dir: Path, update_dict: dict):
         json.dump(config, f)
 
 
+@contextmanager
+def add_directory(dir: Path):
+    """Context manager that temporarily creates a directory and removes it again if it's empty"""
+    dir.mkdir(exist_ok=True)
+    try:
+        yield
+    finally:
+        try:
+            dir.rmdir()
+        except OSError:
+            # directory not empty
+            pass
+
+
 def check_ask_pre_commit(dir: Path):
     """
     Check if pre-commit hooks are installed and asks to install them
 
     If the user declines, info will be written to `.dso.json` to not ask again in the future.
+
+    Additionally, we respect a `DSO_SKIP_CHECK_ASK_PRE_COMMIT` environment variable. If it is set
+    to anything that evaluates as `True`, we skip the check and question altogether. This was a
+    requirement introduced by the R API package: https://github.com/Boehringer-Ingelheim/dso/issues/50.
     """
+    if environ.get("DSO_SKIP_CHECK_ASK_PRE_COMMIT", None):
+        return
     config = _read_dot_dso_json(dir)
     if config.get("check_ask_pre_commit", True):
         project_root = get_project_root(dir)