+++ CHANGELOG.md
@@ -8,6 +8,154 @@ and this project adheres to [Semantic Versioning][].
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## v0.12.2
+
+### Fixes
+
+-   Empty folders were not created as expected in stage templates ([#107](https://github.com/Boehringer-Ingelheim/dso/pull/107))
+
+### Documentation
+
+-   Restructure into user guides and tutorials section ([#104](https://github.com/Boehringer-Ingelheim/dso/pull/104))
+-   Add tutorial on "how to work on existing DSO project" ([#104](https://github.com/Boehringer-Ingelheim/dso/pull/104))
+-   Add tutorial on "how to convert an existing project into DSO format" ([#104](https://github.com/Boehringer-Ingelheim/dso/pull/104))
+
+## v0.12.1
+
+### Fixes
+
+-   Temporarily disable linting, because of its slow speed with only one rule implemented. See #70, #5, and #66
+    on GitHub for more information ([#103](https://github.com/Boehringer-Ingelheim/dso/pull/103)).
+
+## v0.12.0
+
+### Migration advice
+
+The default pre-commit configuration has been reworked. To update it, navigate to the root of your project. Then run
+
+```bash
+rm .pre-commit-config.yaml
+dso init .
+```
+
+dso init will re-add all files from the project template that are missing from your project. Existing files will not be touched.
+
+### Template updates
+
+-   Update `.pre-commit-config.yaml`, removing unnecessary hooks ([#99](https://github.com/Boehringer-Ingelheim/dso/pull/99)).
+
+### New features
+
+-   Add `dso pull` command, a wrapper around `dso compile-config` + `dvc pull` ([#99](https://github.com/Boehringer-Ingelheim/dso/pull/99))
+-   Add templates for Python stages (`quarto_py`, `quarto_ipynb`) ([#98](https://github.com/Boehringer-Ingelheim/dso/pull/98)).
+
+### Documentation
+
+-   Update documentation, finalizing the most important sections of the user guide.
+
+## v0.11.0
+
+### Template updates
+
+-   Single `.gitignore` file per stage. Content of input/output/report folders is ignored. These folders
+    do not contain a separate `.gitignore` anymore. This means empty folders won't be tracked by git, but
+    this solves issues with dvc refusing to track the output folder because it is already partly tracked by git ([#73](https://github.com/Boehringer-Ingelheim/dso/pull/73)).
+
+### Fixes
+
+-   Do not change the configuration of the root logger, only the `dso` logger. Changing the root logger
+    had side-effects on other libraries when importing `dso` in Python ([#80](https://github.com/Boehringer-Ingelheim/dso/pull/80)).
+
+### New features
+
+-   Paths in `params.in.yaml` files declared with `!path` can now be compiled to absolute instead of relative paths ([#78](https://github.com/Boehringer-Ingelheim/dso/pull/78)).
+-   Python API that mirrors `dso-r` functionality (e.g. to be used from Jupyter notebooks) ([#30](https://github.com/Boehringer-Ingelheim/dso/pull/30))
+-   `dso exec quarto` automatically creates an `output` directory in the stage if it doesn't exist. If it doesn't contain any file,
+    it will be removed again after completion ([#73](https://github.com/Boehringer-Ingelheim/dso/pull/73)).
+
+### Documentation
+
+-   Various documentation updates, working towards the first public version of the docs.
+
+### Chore
+
+-   Refactor CLI into separate module ([#30](https://github.com/Boehringer-Ingelheim/dso/pull/30))
+-   Defer imports in CLI until they are actually needed to speed up CLI ([#30](https://github.com/Boehringer-Ingelheim/dso/pull/30))
+-   Make all modules explicitly private that are not part of the public API ([#30](https://github.com/Boehringer-Ingelheim/dso/pull/30))
+-   Relicense the package as LGPL-3.0-or-later, with a more permissive exception for the templates ([#76](https://github.com/Boehringer-Ingelheim/dso/pull/76))
+
+## v0.10.1
+
+### Fixes
+
+-   Take comments into account when linting for `DSO001` ([#69](https://github.com/Boehringer-Ingelheim/dso/pull/69))
+-   Make it possible to override watermarks/disclaimers with a simple `null` ([#69](https://github.com/Boehringer-Ingelheim/dso/pull/69)).
+-   Compile _all_ configs on `dso repro`, not just the ones relvant to the specified stage. This is required because we don't
+    know which other stages dvc might compile ([#69](https://github.com/Boehringer-Ingelheim/dso/pull/69)).
+-   Make `get-config` compatible with dvc matrix stages ([#69](https://github.com/Boehringer-Ingelheim/dso/pull/69)).
+
+### Template updates
+
+-   Do not ignore the `.gitignore` files in output/report directories of template ([#63](https://github.com/Boehringer-Ingelheim/dso/pull/63))
+-   Update `.pre-commit-config.yaml` for pre-commit 4.x ([#63](https://github.com/Boehringer-Ingelheim/dso/pull/63))
+
+## v0.10.0
+
+### Template updates
+
+-   Improve instruction text in quarto template to get users started more quickly ([#40](https://github.com/Boehringer-Ingelheim/dso/pull/40))
+-   Add `.gitignore` catch-all clauses for `output` and `report` folders in stages to not pick up data and repots being tracked via git ([#46](https://github.com/Boehringer-Ingelheim/dso/issues/46)).
+-   Every dso project is now also a [uv project](https://docs.astral.sh/uv/concepts/projects/#building-projects) declaring Python dependencies in `pyproject.toml`. This makes it possible to
+    use a different dso version per project and makes it easy to work with virtual Python environments ([#52](https://github.com/Boehringer-Ingelheim/dso/pull/52))
+-   bash templates now include `-euo pipefail` settings, ensuring that stages fail early and return a nonzero error code if something failed ([#59](https://github.com/Boehringer-Ingelheim/dso/pull/59)).
+
+### Fixes
+
+-   Remove vendored `hiyapyco` code since required changes were released upstream in v0.7.0 ([#45](https://github.com/Boehringer-Ingelheim/dso/pull/45)).
+-   `None` values in `params.in.yaml` can now be used to override anything, e.g. to disable watermarking only in a specific stage ([#45](https://github.com/Boehringer-Ingelheim/dso/pull/45)).
+-   Clean up existing `*.rmarkdown` files in quarto stage before running `quarto render`. This fixes issues with re-running quarto stages that failed in the previous attempt ([#57](https://github.com/Boehringer-Ingelheim/dso/pull/57)).
+-   DSO now respects a `DSO_SKIP_CHECK_ASK_PRE_COMMIT` environment variable. If it is set
+    to anything that evaluates as `True`, we skip the check if pre-commit is installed. This was a
+    requirement introduced by the R API package, see [#50](https://github.com/Boehringer-Ingelheim/dso/issues/50) ([#58](https://github.com/Boehringer-Ingelheim/dso/pull/58)).
+-   Improve logging for "missing path" warning during `compile-config` ([#59](https://github.com/Boehringer-Ingelheim/dso/pull/59)).
+-   Improve logging for missing parameters in `dvc.yaml` during `get-config` ([#59](https://github.com/Boehringer-Ingelheim/dso/pull/59)).
+-   Make sure internal calls to the dso pandocfilter use the same python and dso version as the parent command. This is important for the upcoming `dso-mgr` feature ([#61](https://github.com/Boehringer-Ingelheim/dso/pull/61)).
+
+## v0.9.0
+
+### New Features
+
+-   `dso watermark` now supports files in PDF format. With this change, quarto reports using the watermark feature can
+    be rendered to PDF, too ([#26](https://github.com/Boehringer-Ingelheim/dso/pull/26)).
+
+### Fixes
+
+-   Fix linting rule DSO001: It is now allowed to specify additional arguments in `read_params()`, e.g. `quiet = TRUE` ([#36](https://github.com/Boehringer-Ingelheim/dso/pull/36)).
+-   It is now possible to use Jinja2 interpolation in combination with `!path` objects ([#36](https://github.com/Boehringer-Ingelheim/dso/pull/36))
+-   Improve error messages when `dso get-config` can't find required input files ([#36](https://github.com/Boehringer-Ingelheim/dso/pull/36))
+
+### Documentation
+
+-   Documentation is now built via sphinx and hosted on GitHub pages: https://boehringer-ingelheim.github.io/dso/ ([#35](https://github.com/Boehringer-Ingelheim/dso/pull/35)).
+
+### Template updates
+
+-   Make instruction comments in quarto template more descriptive ([#33](https://github.com/Boehringer-Ingelheim/dso/pull/33)).
+-   Include `params.yaml` in default project `.gitignore`. We decided to not track `params.yaml` in git anymore
+    since it adds noise during code review and led to merge conflicts in some cases. In the future, a certain
+    `dso` version will be tied to each project, improving reproducibility also without `params.yaml` files.
+
+### Migration advice
+
+-   Add `params.yaml` to your project-level `.gitignore`. Then execute `find -iname "params.yaml" -exec git rm --cached {} \;`
+    to untrack existing `params.yaml` files.
+
+## v0.8.2
+
+### Fixes
+
+-   Fixed params.yaml sorting issue - params.yaml order will be kept intact when using `dso get-config`
+
 ## v0.8.1
 
 ### Fixes
@@ -39,7 +187,6 @@ and this project adheres to [Semantic Versioning][].
     messages are now `debug` messages and not shown by default.
 -   When running `dso repro`, configuration is only compiled once and not recompiled when `dso exec` or `dso get-config`
     is called internally. This reduces runtime and redundant log messages.
-
 
 ## v0.7.0